sensible output no error on rm empty cache ..

+++ images/05.SystemApplications/control/home/engines/scripts/actionators/clear_library_cache.sh
@@ -1,3 +1,8 @@
 #!/bin/sh
-
-rm /home/app/data/v0/library.json
+if test -f /home/app/data/v0/library.json
+ then
+	rm /home/app/data/v0/library.json
+	echo Library cache cleared
+ else
+    echo Library cache was empty
+fi
